change the layout of the explanation
sudo chown root setuid_feature
sudo chmod +s setuid_feature

There are three uids for a process in Linux: REAL uid, EFFECTIVE uid, SAVED uid.
Cond 1. When euid is root, setuid or seteuid can be set to any uid, but there is a side effect, when using setuid(not seteuid),  all of the three can be set to the same uid which is not ROOT, and then the process can't regain ROOT privilege.
Cond 2. When euid is not root, setuid or seteuid can be set to ruid or suid, and only euid is changed.
                         |            seteuid        |               setuid          |
Cond 1. (euid == root)   | set euid to any uid       | set all three uids to any uid |
Cond 2. (euid != root)   | set euid to ruid or suid  | set euid to ruid or suid      |

so, there are 5 setuid or seteuid process there in code, let me classify them:
1. seteuid(600):  Cond 1, set euid to 600
2. setuid(1000):  Cond 2, set euid to 1000
3. setuid(0)   :  Cond 2, set euid to 0(suid)
4. setuid(1000):  Cond 1, set all three uids to 1000
5.bsetuid(0)   :  Cond 2, all three uids is not equal 0, so can't set to 0, failed with ret = 1